+++ R/earcut.R
@@ -1,22 +1,11 @@
-#' Ear cut triangulation
-#' 
-#' Ear cutting triangulation for polygons. 
-#' 
-#' This function builds on `earcut` the package `decido`. 
-#' @name earcut
-#' @export
-#' @importFrom decido earcut
-#' @export earcut
-#' @export 
-#' @param x PATH object to triangulate
-#' @param holes ignored by the PATH method
-#' @param ... 
-#' @examples 
-#' earcut.PATH(PATH(minimal_mesh))
-earcut.PATH <- function(x, holes = NULL, ...) {
+## internal function
+earcut_PATH <- function(x, holes = NULL, ...) {
   path <- silicate::sc_path(x)
   path_link_vertex <- x$path_link_vertex
   vertex <- x$vertex
+  if (nrow(vertex) < 3) stop("need at least 3 coordinates")
+  if (anyNA(vertex$x_)) stop("missing values in x_")
+  if (anyNA(vertex$y_)) stop("missing values in y_")
   ## here we need a new "object", analogous to polygon
   ## can be temporary
   path$hole <- duplicated(path$object)
@@ -28,20 +17,10 @@ earcut.PATH <- function(x, holes = NULL, ...) {
     x <- x[!duplicated(x["vertex_"]), ]
     holes <- which(abs(diff(as.integer(factor(x$path_)))) > 0)
     if (length(holes) < 1) holes <- 0
-    idx <- earcut(x[["x_"]], x[["y_"]], holes, ...)
+    idx <- decido::earcut(x[["x_"]], x[["y_"]], holes, ...)
     out <- match(x$vertex_[idx], vertex$vertex_)
     out
   }))
   tri
 }
 
-#' @name earcut
-#' @export
-earcut.sf <- function(x, ...) {
-  earcut.PATH(PATH(x), ...)
-}
-#' @name earcut
-#' @export
-earcut.sfc <- function(x, ...) {
-  earcut.PATH(PATH(x), ...)
-}